apps/content-engine/pyproject.toml
[project]
name = "iloveberlin-content-engine"
version = "0.1.0"
description = "Automated content pipeline for I♥Berlin"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "httpx>=0.27",
    "feedparser>=6.0",
    "openai>=1.30",
    "apscheduler>=3.10,<4",
    "Pillow>=10.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-slugify>=8.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "structlog>=24.0",
    "bleach>=6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.7",
    "alembic>=1.13",
    "itsdangerous>=2.1",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
